requirements.txt
certifi==2024.6.2
distlib==0.3.8
filelock==3.15.4
pipenv==2024.0.1
platformdirs==4.2.2
python-dotenv==1.0.1
schedule==1.2.2
setuptools==70.1.1
virtualenv==20.26.3
wheel==0.38.4
